+++ tests/testthat/test-fst_ld.R
@@ -236,16 +236,13 @@ test_that("fis bootstrapping",{
   
   # basic
   expect_true("weighted_mean_fis_p" %in% colnames(bs1_res$weighted.means))
+  expect_true("weighted_mean_fis_uCI" %in% colnames(bs1_res$weighted.means))
+  expect_true("weighted_mean_fis_lCI" %in% colnames(bs1_res$weighted.means))
   skip_if_not("weighted_mean_fis_p" %in% colnames(bs1_res$weighted.means))
   
   
+  
   expect_true(is.numeric(unlist(bs1_res$weighted.means$weighted_mean_fis_p)))
-
-  # parallel
-  skip_on_cran();
-  bs1_par <- calc_fis(.internal.data$test_snps, "pop", boot = 10, boot_par = 2)
-  bs1_res <- get.snpR.stats(bs1_par, "pop", "fis")
-  expect_true("weighted_mean_fis_p" %in% colnames(bs1_res$weighted.means))
 
   # complex facets
   bs1_par <- calc_fis(.internal.data$test_snps, c("pop", "fam", "pop.chr", "chr"), boot = 10)
@@ -268,6 +265,12 @@ test_that("fis bootstrapping",{
   expect_true("weighted_mean_fis_p" %in% colnames(bs1_res$weighted.means))
   expect_true(all(bs1_res$weighted.means[bs1_res$weighted.means$facet == ".base",]$subfacet == ".base"))
   expect_true(all(bs1_res$weighted.means[bs1_res$weighted.means$facet == "pop",]$snp.subfacet == ".base"))
+  
+  # parallel
+  skip_on_cran();
+  bs1_par <- calc_fis(.internal.data$test_snps, "pop", boot = 10, boot_par = 2)
+  bs1_res <- get.snpR.stats(bs1_par, "pop", "fis")
+  expect_true("weighted_mean_fis_p" %in% colnames(bs1_res$weighted.means))
 })
 
 test_that("global fst",{